# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flatmates Shared Expenses API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups the current user belongs to
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details with current members
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group name
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List all members (including past) with join/leave dates
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to the group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /groups/{groupId}/members/{membershipId}:
    patch:
      operationId: updateGroupMember
      tags: [groups]
      summary: Update membership (set leave date)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: membershipId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a membership record
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: membershipId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Expenses ───────────────────────────────────────────────────────────────
...
  /groups/{groupId}/settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      summary: List settlements for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Record a settlement payment
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
  /settlements/{settlementId}:
    delete:
      operationId: deleteSettlement
      tags: [settlements]
      summary: Delete a settlement
      parameters:
        - name: settlementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Balances ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Group:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
    GroupDetail:
      type: object
      required: [id, name, createdAt, members]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        memberNames:
          type: array
          items:
            type: string
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
    Membership:
      type: object
      required: [id, groupId, userId, userName, joinedAt, isActive]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        joinedAt:
          type: string
        leftAt:
          type: ["string", "null"]
        isActive:
          type: boolean
    MembershipInput:
      type: object
      required: [userId, joinedAt]
      properties:
        userId:
          type: integer
        joinedAt:
          type: string
        userName:
          type: string
    MembershipUpdate:
      type: object
      properties:
        leftAt:
          type: ["string", "null"]
        joinedAt:
          type: string

    # ── Expenses ──────────────────────────────────────────────────────────────
...
    Settlement:
      type: object
      required: [id, groupId, fromUserId, fromUserName, toUserId, toUserName, amount, date, createdAt]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        fromUserId:
          type: integer
        fromUserName:
          type: string
        toUserId:
          type: integer
        toUserName:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SettlementInput:
      type: object
      required: [fromUserId, toUserId, amount, date]
      properties:
        fromUserId:
          type: integer
        toUserId:
          type: integer
        amount:
          type: number
        date:
          type: string
        notes:
          type: string

    # ── Balances ──────────────────────────────────────────────────────────────
...