# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  name: string;
  memberNames?: string[];
}