# lib/api-zod/src/generated/types/membership.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */

export interface Membership {
  id: number;
  groupId: number;
  userId: number;
  userName: string;
  userEmail?: string;
  joinedAt: string;
  /** @nullable */
  leftAt?: string | null;
  isActive: boolean;
}