# lib/api-zod/src/generated/types/membershipInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flatmates Shared Expenses API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipInput {
  userId: number;
  joinedAt: string;
  userName?: string;
}